requirements.txt
alpaca-py>=0.21.0
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.24.0
rich>=13.0.0
pytz>=2024.1
yfinance>=0.2.36
xgboost>=2.0.0
scikit-learn>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
statsmodels>=0.14.0
aiosqlite>=0.19.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
scipy>=1.11.0
pykalman>=0.9.7
